Some further work.
As soon as I get setting kernel environment and some kernel stub up, I will commit
some generated bootfloppy. Tools and instruction on building the current "release"
will appear soon, maybe in next commit.

+++ loader/stage3/screen.h
@@ -23,10 +23,10 @@ namespace Screen
         // most of them will never be used, but who cares
         Console * Print(char);
         Console * Print(char *);
-        Console * Print(long int);
-        Console * Print(long int *);
-        Console * Print(int);
-        Console * Print(int *);
+        Console * Print(unsigned long long int);
+        Console * Print(unsigned long long int *);
+        Console * Print(unsigned int);
+        Console * Print(unsigned int *);
         Console * Print(double);
         Console * Print(double *);
         Console * Print(void *);
